ref: move some util classes out of model/obj
