+++ src/main/java/cam72cam/mod/model/common/util/FaceUtils.java
@@ -1,7 +1,6 @@
 package cam72cam.mod.model.common.util;
 
 import cam72cam.mod.math.Vec3d;
-import cam72cam.mod.model.obj.Buffers;
 import it.unimi.dsi.fastutil.ints.IntArrayList;
 
 import java.util.List;
